requirements.txt
# Core Application Dependencies
pyautogui==0.9.54
pygetwindow==0.0.9
Pillow==10.4.0
requests==2.32.3
opencv-python==4.10.0.84
pywin32==310
numpy==2.1.1
mss==9.0.1
dxcam==0.0.5
python-dotenv==1.0.0

# Modern UI/UX Libraries
PyQt5==5.15.10
qtawesome==1.3.1
qdarkstyle==3.2.3
PyQt5-stubs==5.15.6.0